requirements.txt
cassandra_driver==3.29.1
dnspython==2.6.1
